refactor gethandle to be on the classes
